+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Allow users to select multiple years [#138](https://github.com/azavea/fb-gender-survey-dashboard/pull/138)
+- Format comparative bar charts [#142](https://github.com/azavea/fb-gender-survey-dashboard/pull/142)
 
 ### Changed